requirements.txt
qsdsan>=1.4.2
pandas>=2.2.0
numpy>=1.26.0
joblib>=1.3.0
scipy>=1.11.0
google-cloud-storage>=2.10.0